Refactored to better clean code
Refactor ReconcileController.transactionCompare to recive the model
Inputs and create the model Report, which have the reconcile report.
transactionCompare now send to the view unmatched-report.

Inputs have the private method to convert the file into a list
(getFileTransactions).

Report have all the unmatcheds records with the field outerUnmatcheds,
which use the private method Report.outerJoin to remove all equals
transactions.
The fields Report.unmatched have the reconcile records, which follows
the matched strategy.