Refactor interface ReconcileStrategy
Change ReconcileStrategy to be a interface to helpers classes that
implements the methods isPossibleOfReconciliation and isUnmatched and is
call with Report.reconcileRecords.
That way, we can change the logic of what is a reconcile record.

To demonstrate, we build ReconcileWithUnmatchedId whick acknowledge that
the Transaction.id can be a field that could bring unmatcheds records.