Refactor to not make a second post to unmatched report
Change the form on comparison-result to a button with the onclick
function showTable.

Add the script showTable on the form.jsp.

Change the style to the div of unmatched-report.jsp of 'display' to
'visible'. The onclick function doesn't work with 'display'.

ReconcileController.unmatchedReport was delete because doesn't has no
need anymore. The field report was delete too, beacause it was a
workaround to show on the view.