Refactor with Shotgun Surgery
To simplify the constructor of Report, we did a shotgun surgery with the
duplicate code of the methods Report.reconcileRecords and
Report.subtract.
To this, we create two new classes:

SubtracctionsOfLists, with the responsibility of subtract one list from
another, and save the result within theirs fields; l1MinusL2 and
l2MinusL1.

ReconcileRecords, which order two list by the ReconcileStrategy passed
by his constructor.

With this, we can reference the two values needed by one variable.

Refactor ReportTest and ReconcileController two used the new
architecture.

Create a new ReconcileStrategy: ReconcileWithUnmatchedWallet.